+++ packages/blockchain-bindings/src/index.ts
@@ -1,3 +1,4 @@
 export * from './contracts';
 export * from './types';
+export * from './TypedData';
 export * from './utils';