+++ packages/react/src/environments.ts
@@ -1,3 +1,4 @@
+import { ContentInsightMatcher, demoSnapshotPoll, snapshotPoll } from '@lens-protocol/api-bindings';
 import * as GatedEnvironments from '@lens-protocol/gated-content/environments';
 import { ChainType, Url } from '@lens-protocol/shared-kernel';
 
@@ -5,6 +6,17 @@ import { ChainConfigRegistry, goerli, mainnet, mumbai, polygon } from './chains'
 import { TransactionObserverTimings } from './transactions/infrastructure/TransactionObserver';
 
 export type { TransactionObserverTimings };
+
+/**
+ * The Snapshot.org integration configuration
+ *
+ * @internal
+ */
+export type SnapshotConfig = {
+  hub: Url;
+  sequencer: Url;
+  matcher: ContentInsightMatcher;
+};
 
 /**
  * A function that resolves a profile handle to a fully qualified profile handle
@@ -24,6 +36,7 @@ export type EnvironmentConfig = {
   chains: ChainConfigRegistry;
   timings: TransactionObserverTimings;
   handleResolver: ProfileHandleResolver;
+  snapshot: SnapshotConfig;
   gated: GatedEnvironments.EnvironmentConfig;
 };
 
@@ -50,9 +63,13 @@ export const production: EnvironmentConfig = {
     maxMiningWaitTime: 60000,
   },
   handleResolver: (handle) => `${handle}.lens`,
+  snapshot: {
+    hub: 'https://hub.snapshot.org',
+    matcher: snapshotPoll,
+    sequencer: 'https://seq.snapshot.org',
+  },
   gated: GatedEnvironments.production,
 };
-
 /**
  * The development environment configuration
  *
@@ -76,6 +93,11 @@ export const development: EnvironmentConfig = {
     maxMiningWaitTime: 120000,
   },
   handleResolver: (handle) => `${handle}.test`,
+  snapshot: {
+    hub: 'https://testnet.snapshot.org',
+    matcher: demoSnapshotPoll,
+    sequencer: 'https://testnet.seq.snapshot.org',
+  },
   gated: GatedEnvironments.development,
 };
 
@@ -103,6 +125,11 @@ export const sandbox: EnvironmentConfig = {
     maxMiningWaitTime: 120000,
   },
   handleResolver: (handle) => `${handle}.test`,
+  snapshot: {
+    hub: 'https://testnet.snapshot.org',
+    matcher: demoSnapshotPoll,
+    sequencer: 'https://testnet.seq.snapshot.org',
+  },
   gated: GatedEnvironments.sandbox,
 };