+++ packages/react/src/publication/__tests__/usePublications.spec.ts
@@ -1,6 +1,6 @@
 import { activeProfileIdentifierVar, AnyPublication } from '@lens-protocol/api-bindings';
 import {
-  createMockApolloClientWithMultipleResponses,
+  mockLensApolloClient,
   mockPostFragment,
   createGetPublicationsMockedResponse,
   mockSources,
@@ -23,7 +23,7 @@ function setupTestScenario({
   return renderHookWithMocks(() => usePublications(args), {
     mocks: {
       sources,
-      apolloClient: createMockApolloClientWithMultipleResponses([
+      apolloClient: mockLensApolloClient([
         createGetPublicationsMockedResponse({
           variables: {
             ...args,
@@ -51,7 +51,7 @@ describe(`Given the ${usePublications.name} hook`, () => {
       const { result } = setupTestScenario({ profileId, result: publications });
 
       await waitFor(() => expect(result.current.loading).toBeFalsy());
-      expect(result.current.data).toEqual(publications);
+      expect(result.current.data).toMatchObject(publications);
     });
   });
 
@@ -70,7 +70,7 @@ describe(`Given the ${usePublications.name} hook`, () => {
       });
 
       await waitFor(() => expect(result.current.loading).toBeFalsy());
-      expect(result.current.data).toEqual(publications);
+      expect(result.current.data).toMatchObject(publications);
     });
 
     it('should always allow to specify the "observerId" on a per-call basis', async () => {
@@ -84,7 +84,7 @@ describe(`Given the ${usePublications.name} hook`, () => {
       });
 
       await waitFor(() => expect(result.current.loading).toBeFalsy());
-      expect(result.current.data).toEqual(publications);
+      expect(result.current.data).toMatchObject(publications);
     });
   });
 });