+++ packages/react/src/transactions/adapters/__tests__/PublicationCacheManager.spec.ts
@@ -1,27 +1,24 @@
-import { AnyPublication } from '@lens-protocol/api-bindings';
+import { AnyPublication, FragmentPublication } from '@lens-protocol/api-bindings';
 import {
-  createMockApolloCache,
+  mockLensCache,
   mockPostFragment,
   mockCommentFragment,
   mockMirrorFragment,
 } from '@lens-protocol/api-bindings/mocks';
+import { mockPublicationId } from '@lens-protocol/domain/mocks';
 
-import {
-  FragmentNotFoundError,
-  resolveFragmentDoc,
-  PublicationCacheManager,
-} from '../PublicationCacheManager';
+import { PublicationCacheManager } from '../PublicationCacheManager';
 
 function setupTestScenario({ publication }: { publication: AnyPublication }) {
-  const apolloCache = createMockApolloCache();
+  const apolloCache = mockLensCache();
 
   apolloCache.writeFragment({
     id: apolloCache.identify({
       __typename: publication.__typename,
       id: publication.id,
     }),
-    fragment: resolveFragmentDoc(publication),
-    fragmentName: publication.__typename,
+    fragment: FragmentPublication,
+    fragmentName: 'Publication',
     data: publication,
   });
 
@@ -34,8 +31,8 @@ function setupTestScenario({ publication }: { publication: AnyPublication }) {
           __typename: publication.__typename,
           id: publication.id,
         }),
-        fragment: resolveFragmentDoc(publication),
-        fragmentName: publication.__typename,
+        fragment: FragmentPublication,
+        fragmentName: 'Publication',
       });
     },
   };
@@ -43,80 +40,38 @@ function setupTestScenario({ publication }: { publication: AnyPublication }) {
 
 describe(`Given the ${PublicationCacheManager.name}`, () => {
   describe(`when "${PublicationCacheManager.prototype.update.name}" method is invoked`, () => {
-    it(`should be able to update post cache fragments`, async () => {
-      const post = mockPostFragment({
+    it.each([
+      mockPostFragment({
         hidden: false,
-      });
+      }),
+      mockCommentFragment({
+        hidden: false,
+      }),
+      mockMirrorFragment({
+        hidden: false,
+      }),
+    ])(`should be able to update $__typename cache fragments`, async (publication) => {
+      const scenario = setupTestScenario({ publication });
 
-      const scenario = setupTestScenario({
-        publication: post,
-      });
-
-      scenario.cacheManager.update(post.id, (data) => ({
+      scenario.cacheManager.update(publication.id, (data) => ({
         ...data,
         hidden: true,
       }));
 
-      expect(scenario.updatedPublicationFragment).toEqual(
-        expect.objectContaining({
-          hidden: true,
-        }),
-      );
-    });
-
-    it(`should be able to update comment cache fragments`, async () => {
-      const comment = mockCommentFragment({
-        hidden: false,
-      });
-
-      const scenario = setupTestScenario({
-        publication: comment,
-      });
-
-      scenario.cacheManager.update(comment.id, (data) => ({
-        ...data,
+      expect(scenario.updatedPublicationFragment).toMatchObject({
         hidden: true,
-      }));
-
-      expect(scenario.updatedPublicationFragment).toEqual(
-        expect.objectContaining({
-          hidden: true,
-        }),
-      );
+      });
     });
 
-    it(`should be able to update mirror cache fragments`, async () => {
-      const mirror = mockMirrorFragment({
-        hidden: false,
-      });
-
-      const scenario = setupTestScenario({
-        publication: mirror,
-      });
-
-      scenario.cacheManager.update(mirror.id, (data) => ({
-        ...data,
-        hidden: true,
-      }));
-
-      expect(scenario.updatedPublicationFragment).toEqual(
-        expect.objectContaining({
-          hidden: true,
-        }),
-      );
-    });
-
-    it(`should return failure if the specified cache item is not present`, async () => {
+    it(`should not call the update callback if the cache item is not present`, async () => {
       const scenario = setupTestScenario({
         publication: mockPostFragment(),
       });
+      const updateSpy = jest.fn();
 
-      const result = scenario.cacheManager.update(mockMirrorFragment().id, (data) => ({
-        ...data,
-        hidden: true,
-      }));
+      scenario.cacheManager.update(mockPublicationId(), updateSpy);
 
-      expect(() => result.unwrap()).toThrow(FragmentNotFoundError);
+      expect(updateSpy).not.toHaveBeenCalled();
     });
   });
 });