+++ packages/api-bindings/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @lens-protocol/api-bindings
 
+## 0.9.0-next.1
+
+### Minor Changes
+
+- 225f0fa7: **Added** `usePollDetail` and `usePollVote` hooks
+
+### Patch Changes
+
+- 3025d56a: allow self funded fallback for proxy actions
+- 97ecba69: **Fixed** cache redirects for Publication and Profile
+- Updated dependencies [3025d56a]
+- Updated dependencies [225f0fa7]
+- Updated dependencies [422c627e]
+  - @lens-protocol/domain@0.9.0-next.1
+  - @lens-protocol/shared-kernel@0.9.0-next.0
+
 ## 0.8.1-next.0
 
 ### Patch Changes