+++ packages/gated-content/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @lens-protocol/gated-content
 
+## 0.3.0-next.2
+
+### Patch Changes
+
+- Updated dependencies [3025d56a]
+- Updated dependencies [225f0fa7]
+- Updated dependencies [422c627e]
+- Updated dependencies [97ecba69]
+  - @lens-protocol/api-bindings@0.9.0-next.1
+  - @lens-protocol/domain@0.9.0-next.1
+  - @lens-protocol/shared-kernel@0.9.0-next.0
+  - @lens-protocol/storage@0.7.3-next.0
+
 ## 0.3.0-next.1
 
 ### Minor Changes